Adding NER classification files
Got NER working and decided to merge it in. You can test the NER by
running the NERDemo in the NER package. If you run it with args the
first arg is the path of the classifer you want to use. The second is
the file you want to read from.